+++ validators/filenameValidator.js
@@ -5,29 +5,29 @@ const {
   matchedData,
 } = require("express-validator");
 
-const folderSchema = {
-  folder_name: {
+const filenameSchema = {
+  upload_files: {
     trim: true,
     isEmpty: {
       negated: true,
       bail: true,
-      errorMessage: "Folder name cannot be empty.",
+      errorMessage: "File name cannot be empty.",
     },
     isLength: {
       options: { max: 10 },
-      errorMessage: "Folder name cannot exceed 10 characters in length.",
+      errorMessage: "File name cannot exceed 10 characters in length.",
     },
     escape: true,
   },
 };
 
-const validateFolder = (view) => {
+const validateFilename = (view) => {
   return asyncHandler(async (req, res, next) => {
-    await checkSchema(folderSchema, ["body"]).run(req);
+    console.log("testing...");
+    await checkSchema(filenameSchema, ["body"]).run(req);
     const errors = validationResult(req);
     const inputs = matchedData(req, { onlyValidData: false });
-    console.log(errors);
-
+    console.log("errors:", errors);
     if (!errors.isEmpty()) {
       return res.status(422).render(view, {
         errors: errors.mapped(),
@@ -39,4 +39,4 @@ const validateFolder = (view) => {
   });
 };
 
-module.exports = validateFolder;
+module.exports = validateFilename;